# api/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Case eview Documents API",
        "description": "Search pre-existing case documents",
        "version": "1.0.0",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "name": "API Support",
            "email": "IET@cica.gov.uk"
        }
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "tags": [
        {
            "name": "Search"
        }
    ],
    "paths": {
        "/search/{query}/{pageNumber}/{itemsPerPage}": {
            "get": {
                "tags": ["Search"],
                "summary": "Search documents for a given case",
                "description": "Gets all matching occurrences of text from case documents",
                "operationId": "getSearchResults",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/query"
                    },
                    {
                        "$ref": "#/components/parameters/pageNumber"
                    },
                    {
                        "$ref": "#/components/parameters/itemsPerPage"
                    },
                    {
                        "$ref": "#/components/parameters/onBehalfOf"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "json-schemas/api/_search_{query}/get/res/200.json"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "type": "search-results",
                                            "id": "08dbd0d2-6cee-49e5-a55f-35b31ac4aa9e",
                                            "attributes": {
                                                "query": "gabapentin"
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/documents": {
            "get": {
                "tags": ["Documents"],
                "summary": "Get all documents for a given case",
                "description": "Gets all physical documents attached to the case reference number",
                "operationId": "getDocuments",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "json-schemas/api/_documents/get/res/200.json"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "type": "documents",
                                            "id": "08dbd0d2-6cee-49e5-a55f-35b31ac4aa9e",
                                            "attributes": {
                                                
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/documents/{documentId}": {
            "get": {
                "tags": ["Documents"],
                "summary": "Get document for a given case",
                "description": "Gets a single physical document attached to the case reference number",
                "operationId": "getDocument",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "json-schemas/api/_documents_{documentId}/get/res/200.json"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "type": "documents",
                                            "id": "08dbd0d2-6cee-49e5-a55f-35b31ac4aa9e",
                                            "attributes": {
                                                
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "parameters": {
            "query": {
                "name": "query",
                "in": "path",
                "required": true,
                "description": "The query that results in the specific search results instance. Format UUIDv4.",
                "schema": {
                    "type": "string",
                    "pattern": "^[a-zA-Z0-9]*$",
                    "maxLength": 200
                },
                "example": "this is a valid search query"
            },
            "pageNumber": {
                "name": "pageNumber",
                "in": "path",
                "required": true,
                "description": "Page number of search results",
                "schema": {
                    "type": "integer",
                    "min": 1
                },
                "example": 2
            },
            "itemsPerPage": {
                "name": "itemsPerPage",
                "in": "path",
                "required": true,
                "description": "Number of items per page of search results",
                "schema": {
                    "type": "integer",
                    "min": 1,
                    "max": 100
                },
                "example": 2
            },
            "onBehalfOf": {
                "name": "On-Behalf-Of",
                "in": "header",
                "required": true,
                "description": "Specify the owner of the resource.",
                "schema": {
                    "$ref": "./json-schemas/models/definitions/case-reference-number.json"
                },
                "example": "urn:uuid:f81d4fae-7dec-11d0-a765-00a0c91e6bf6"
            }
        },
        "responses": {
            "BadRequest": {
                "description": "There is an issue with the request",
                "content": {
                    "application/vnd.api+json": {
                        "schema": {
                            "$ref": "./json-schemas/api/responses/400.json"
                        },
                        "example": {
                            "errors": [
                                {
                                    "status": 400,
                                    "title": "400 Bad Request",
                                    "detail": "Request JSON is malformed"
                                }
                            ]
                        }
                    }
                }
            },
            "UnauthorizedError": {
                "description": "Access token is missing or invalid",
                "content": {
                    "application/vnd.api+json": {
                        "schema": {
                            "$ref": "./json-schemas/api/responses/401.json"
                        },
                        "example": {
                            "errors": [
                                {
                                    "status": 401,
                                    "detail": "No authorization token was found"
                                }
                            ]
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "The JWT doesn't permit access to this endpoint",
                "content": {
                    "application/vnd.api+json": {
                        "schema": {
                            "$ref": "./json-schemas/api/responses/403.json"
                        },
                        "example": {
                            "errors": [
                                {
                                    "status": 403,
                                    "title": "403 Forbidden",
                                    "detail": "Insufficient scope"
                                }
                            ]
                        }
                    }
                }
            },
            "NotFound": {
                "description": "The specified resource was not found",
                "content": {
                    "application/vnd.api+json": {
                        "schema": {
                            "$ref": "./json-schemas/api/responses/404.json"
                        },
                        "example": {
                            "errors": [
                                {
                                    "status": 404,
                                    "title": "404 Not Found",
                                    "detail": "Resource /api/v1/questionnaires/2d7caf89-2c72-469f-b19d-17f2a22270b6/sections/answers does not exist"
                                }
                            ]
                        }
                    }
                }
            },
            "Conflict": {
                "description": "Conflict",
                "content": {
                    "application/vnd.api+json": {
                        "schema": {
                            "$ref": "./json-schemas/api/responses/409.json"
                        },
                        "example": {
                            "errors": [
                                {
                                    "status": 409,
                                    "title": "409 conflict",
                                    "detail": "Submission resource with ID \"68653be7-877f-4106-b91e-4ba8dac883f3\" already exists"
                                }
                            ]
                        }
                    }
                }
            }
        }
    }
}